[monarch][hyperactor] Fix race in proc shutdown
This diff fixes a (highly unlikely, but still real) race during `Proc::destroy_and_wait` where we might call `unwrap()` on an `InstanceCell`'s actor task handle before the handle is actually set, causing a panic. It replaces `unwrap()` with `wait()` so that we are guaranteed to have an actor task handle. See the comment in code for why this should be safe to do (and not hang forever).

Differential Revision: [D82355158](https://our.internmc.facebook.com/intern/diff/D82355158/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82355158/)!

[ghstack-poisoned]