[hyperactor] deprecate 'caps', and introduce simpler "context" module
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1146

The capabilities were unnecessarily fine-grained. In practice, they were used to carry actor contexts. Now that we also have formal support for managing actor instances manually (e.g., to provide a "client actor"), we can simplify this system.

This diff introduces the `hyperactor::context` module, which provides two types: `context::Mailbox` and `context::Actor`, denoting the type of context. The vast majority of call sites should simply use `context::Actor`. You can read this as "this method requires an actor context in which to be called".

We implement the relevant capabilities for both context types, so that we can gradually transition. The goal is then to delete the `cap` module entirely.

This change also includes a change to the `v1` mesh APIs to use `context::Actor`. This will then allow us to cleanly associate mesh ownership with the calling actor.
ghstack-source-id: 309559328
@exported-using-ghexport

Differential Revision: [D82049326](https://our.internmc.facebook.com/intern/diff/D82049326/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82049326/)!