Update base for Update on "[monarch] switch over (mostly) to the new context module"
This was a little more involved than I thought it would be. Primarily, this was driven by porting the macros to use `&impl context::Actor` instead of capabilities. We then converted everything downstream of this. The one exception is the RDMA manager, which is Rust code, invoking macro-genned code, dispatched through Python. I left a `_deprecated` version of each client method to use the old caps (which can be provided by `Mailbox`). Once slurye's work plumbing through Rust Context/Instance is done, we can switch these over as well, and remove the legacy codegen.

Some other casualties include some mock-oriented tests in multiprocess (but we're deprecating this crate anyway).

Differential Revision: [D82248674](https://our.internmc.facebook.com/intern/diff/D82248674/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82248674/)!

[ghstack-poisoned]