Update on "[monarch] Reimplement python bindings for `Instance` and `Context` in terms of the actual underlying instance"
Now that `Instance<A>` has only cloneable state, we can implement the python `Instance` and `Context` objects in terms of `Instance<A>` (rather than just the mailbox and actor id). This is necessary because of the new `context::Actor` trait that replaces the caps system -- we need access to the actual `Instance<A>` in python so that we can spawn new proc meshes/actor meshes. Unfortunately pyo3 doesn't support generic types in structs, so the `PyInstance` struct actually contains a new `ErasedInstance`. This contains all of the same state as `Instance<A>` but with `Arc<Ports<A>>` as `Arc<dyn Any + Send + Sync>` and can be downcast to the concrete `Instance<A>`. I originally wanted to just explicitly use `Instance<PythonActor>` in `PyInstance` but this doesn't work because the global root client is an `Instance<()>`.

Differential Revision: [D82476081](https://our.internmc.facebook.com/intern/diff/D82476081/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82476081/)!

[ghstack-poisoned]