[monarch] Reimplement python bindings for `Instance` and `Context` in terms of the actual underlying instance
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1222

Now that `Instance<A>` has only cloneable state, we can implement the python `Instance` and `Context` objects in terms of `Instance<A>` (rather than just the mailbox and actor id). This is necessary because of the new `context::Actor` trait that replaces the caps system -- we need access to the actual `Instance<A>` in python so that we can spawn new proc meshes/actor meshes. Unfortunately pyo3 doesn't support generic types in structs, so the `PyInstance` struct actually contains a new struct called `ContextInstance`, which multiplexes `Instance<PythonActor>` and `Instance<()>` (for the root client context).
ghstack-source-id: 310134496
@exported-using-ghexport

Differential Revision: [D82476081](https://our.internmc.facebook.com/intern/diff/D82476081/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82476081/)!