[hyperactor] mesh: drop Mesh::freeze() in favor of implementing `Deref`
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1269

Meshes are now dereferencable to MeshRefs.

The idea is that meshes act like smart pointers: they point to a particular state of the mesh, and dereferencing them accesses that current state. After reconfiguration, dereferencing will return a different mesh ref. This is similar to how a smart pointer's pointee can be updated.

This also means that we can share MeshRefs while relying on Rust's borrowing rules to prevent reconfiguration of meshes that have extant references. In other words, the user has to explicitly clone out a reference in order to free its parent mesh -- so we get good ergonomics, and also safety by default.
ghstack-source-id: 310826848
@exported-using-ghexport

Differential Revision: [D82748548](https://our.internmc.facebook.com/intern/diff/D82748548/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82748548/)!