Update base for Update on "[monarch][hyperactor] Fix another race condition in proc shutdown"
This fixes the tests in T238124137. The problem is that if `proc.destroy_and_wait` is called from inside an actor, then after aborting that task's handle, `destroy_and_wait` never makes it past this await point: https://fburl.com/code/xzgtv3mj. This wasn't a problem before because the await point was added recently. The solution is to delay aborting the actor task handle that `proc.destroy_and_wait` was called from.

Differential Revision: [D82558878](https://our.internmc.facebook.com/intern/diff/D82558878/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82558878/)!

[ghstack-poisoned]