[hyperactor] attrs: require that values implement AttrValue
We introduce `AttrValue`, which must be implemented by all value types in attribute dictionary. AttrValue 1) gathers all the requirements (Send, Serialize, Named, ...) of attribute values, and 2) introduces `AttrValue::display` and `AttrValue::parse`, to be used when attribute values are presented to humans.

In most cases, `display` and `parse` can be derived from the type's `ToString` and `FromStr` implementations. The change includes a derive macro to bridge these.

Differential Revision: [D82921321](https://our.internmc.facebook.com/intern/diff/D82921321/)

[ghstack-poisoned]