[monarch] [1/n] jobs api: adding trait and local
Beginning the implementation of RFC: https://docs.google.com/document/d/1QJ1WeRYjDElJ_bySgekWVMujdCXc3oIJRB0Cvkjj-UI/edit?tab=t.0

This diff adds:

* The Jobs Trait
* LocalJob

The LocalJob can schedule as if it were a batch job with .apply taking a client_script. This is useful for testing and for consistency with remote batch scheduling but probably won't get much use.

Next up is to pull in the code from monarch.tools so that we can use the Jobs api with the torchx launchers we already support.

Differential Revision: [D83015058](https://our.internmc.facebook.com/intern/diff/D83015058/)

[ghstack-poisoned]