[hyperactor] common Name definition
A common definition of global, unique names in Hyperactor, with:

1) careful, correct-by-construction representation, enforcing namespaces, identifiers, UIDs, etc;
2) rigorous parsing and good error messages.

This name definition will be used to amend the direct-addressed procs to get fully global names in hyperactor, with a careful separation between URL and URI definitions.

This will set us up for 1) the conveniences and simplicities of using URLs (i.e., direct-addressed names); with 2) having globally meaningful names that can be used in control plane applications ("Monarchy"), tooling, etc.

It will also allow us to later more carefully separate concerns of routing and naming in the system.

Differential Revision: [D83117167](https://our.internmc.facebook.com/intern/diff/D83117167/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D83117167/)!

[ghstack-poisoned]