[hyperactor] reinstate type registration for concrete types; display message port types when available
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1306

Since we now derive most `Named` implementations, we can register them uniformly. Howevever registration can only be performed for concrete types, so parameterized types are currently skipped.

We should consider making typehashes "structural", so that we can capture the type parameters independently (they are guaranteed to also be `Named`). This would let us perform dynamic registration. However, serialization would still be an issue in this case.

We also implement type annotations for PortIds: `index<type>`. It is purely a rendering annotation: on parsing, the type is ignored. This is very useful when debugging message passing. Finally, we print the message and header data when

```
[-]E0923 07:47:53.734178 881992 fbcode/monarch/hyperactor/src/mailbox.rs:786] message not delivered, address not routable: port not bound in mailbox; broken link: message was undeliverable, name:undelivered_message_abandoned, actor_name:testclient, actor_id:unix:@FFkssDgaUI9ZCa4iK3KxsUVx,testproc,testclient[0], dest:tcp:127.0.0.1:45627,service,agent[0][16574558974698272213<hyperactor_mesh::resource::CreateOrUpdate<()>>], headers:hyperactor::mailbox::headers::send_timestamp=2025-09-23T14:47:53.202438459+00:00, data:CreateOrUpdate<()>{"name":"[\"test\",[...29 chars] CRC:65b442bd 5b 22 74 65 73 74 22 2c [...21 bytes]","reply":{"phantom":null,"port_id":"[[{\"Dire[...84 chars] CRC:d07205d0 5b 5b 7b 22 44 69 72 65 [...76 bytes]","reducer_spec":null},"spec":null}
```
ghstack-source-id: 311794895

Differential Revision: [D83058961](https://our.internmc.facebook.com/intern/diff/D83058961/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D83058961/)!