[hyperactor] mesh: add a bootstrapping debug facility
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1329

Bootstrapping problems can be especially difficult to debug. We add a specific debug facility that bypasses all other logging, stdio redirection, etc.

This provides the `io::Writer` `bootstrap::Debug`, with the following behavior: if the file "/tmp/monarch-bootstrap-debug.log" exists, then writes are directed to this file; otherwise all writes are discarded. The existence of the file is checked on the first write.

Thus, you can debug bootstrapping problems:

```
$ touch /tmp/monarch-bootstrap-debug.log; tail -f /tmp/monarch-bootstrap-debug.log
...
```
ghstack-source-id: 312192332
@exported-using-ghexport

Differential Revision: [D83212851](https://our.internmc.facebook.com/intern/diff/D83212851/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D83212851/)!