[RFC][monarch] Typed global configuration from python integrated with the Attrs system
This diff is an RFC for introducing a typed global configuration system in python for monarch, integrated with the hyperactor `Attrs` system and the `hyperactor::global::config` module. It introduces the `py_configurable` macro, which automatically generates typed python bindings for instances of `hyperactor::attrs::Key<T>`. Example usage:
```
py_configurable!(py, DEFAULT_TRANSPORT, "DefaultTransport", PyChannelTransport);
```
This means, "for the `DEFAULT_TRANSPORT` attribute, generate a python class with the following stub/functionality:
```
class DefaultTransport:
  @staticmethod
  def get() -> ChannelTransport: ... # Gets the current value of DEFAULT_TRANSPORT from the global config
  @staticmethod
  def set(val: ChannelTransport) -> None: ... # Sets the current value of DEFAULT_TRANSPORT in the global config
```
The leading `py` in the macro invocation is used in the implementation to indicate that `PyChannelTransport` needs to be converted to/from the actual type of the `DEFAULT_TRANSPORT` attribute (which is `hyperactor::channel::ChannelTransport`). If I were better at rust macros (or rust in general) I could probably figure out how to get rid of it.

I also added functionality to allow for a callback function when the config value is set from python. This could be useful when the root client needs to be re-initialized after setting `DEFAULT_TRANSPORT`.

Differential Revision: [D83701581](https://our.internmc.facebook.com/intern/diff/D83701581/)

[ghstack-poisoned]