[hyperactor] `time::Alarm`
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1404

This introduces `time::Alarm`. It behaves like a (concurrent) alarm: its owner can arm, rearm, and disarm the alarm; an arbitrary number of sleepers can be awoken by the alarm.

Each sleeper is awoken only once for each time the alarm is armed.

This is useful when implementing timed coordination between concurrent tasks, such as is the case when flushing buffers, etc.

This will be used to implement timed flushing (i.e., Nagle's algorithm) for port reducers.
ghstack-source-id: 313884428

Differential Revision: [D83768513](https://our.internmc.facebook.com/intern/diff/D83768513/)