[monarch] Better handling of alloc-specific default bootstrap command
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1588

Previously I had added some custom logic for `ProcessAllocator` inside `HostMesh.allocate_nonblocking` that specified a default bootstrap command if the user didn't specify one. Turns out, this is also needed for `MockMastAllocator`. This diff adds a way for allocators to specify a default bootstrap command for v1 to use if the user doesn't specify one.
ghstack-source-id: 317052639
@exported-using-ghexport

Differential Revision: [D84884566](https://our.internmc.facebook.com/intern/diff/D84884566/)