[monarch] Add configurable stacktrace on unawaited pytokio PythonTask, and make an unawaited error crash the process
This diff adds the environment variable `MONARCH_HYPERACTOR_UNAWAITED_PYTOKIO_TRACEBACK`, which is 0 by default. When it is set to 1, errors in unawaited pytokio `PythonTask` instances will print a traceback indicating where the `PythonTask` was created in addition to the error itself. This should only be used for development/debugging, because it adds overhead to compute the traceback for each new `PythonTask`.

This diff also makes it so that errors in unawaited pytokio tasks crash the process instead of just printing the panic trace. This makes it more consistent with the monarch philosophy of "any unhandled failure should crash".

Differential Revision: [D85614449](https://our.internmc.facebook.com/intern/diff/D85614449/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D85614449/)!

[ghstack-poisoned]