[monarch] Clean future cancellation handling
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1687

The implementation of `monarch.actor.Future` uses `loop.call_soon_threadsafe` to set a value/exception on an underlying asyncio future. But if the caller cancelled the asyncio future before `call_soon_threadsafe` executes the callback, then `asyncio.exceptions.InvalidStateError` will be logged, which is noisy and misleading.

This diff solves the issue by checking whether the future is already resolved/cancelled from inside the callback.
ghstack-source-id: 319407775
@exported-using-ghexport

Differential Revision: [D85693875](https://our.internmc.facebook.com/intern/diff/D85693875/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D85693875/)!